refactor(-): :hammer: move models protocolo padrao
