+++ sme_sigpae_api/dieta_especial/__tests__/commands/test_unifica_dietas_tipo_b.py
@@ -16,7 +16,7 @@ from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadas,
     LogQuantidadeDietasAutorizadasCEI,
 )
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.escola.fixtures.factories.escola_factory import (
