+++ sme_sigpae_api/dieta_especial/__tests__/tasks/test_gerar_logs_ultimo_dia_letivo.py
@@ -7,7 +7,7 @@ from sme_sigpae_api.dados_comuns.models import LogSolicitacoesUsuario
 from sme_sigpae_api.dieta_especial.logs_models.models import (
     LogQuantidadeDietasAutorizadas,
 )
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.dieta_especial.tasks import gera_logs_dietas_especiais_diariamente
