+++ sme_sigpae_api/dieta_especial/__tests__/test_task.py
@@ -10,7 +10,7 @@ from sme_sigpae_api.dados_comuns.models import CentralDeDownload
 from sme_sigpae_api.dieta_especial.api.serializers import (
     SolicitacaoDietaEspecialExportXLSXSerializer,
 )
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.dieta_especial.tasks import (
