+++ sme_sigpae_api/dieta_especial/management/commands/exportar_planilha_relacao_protocolo.py
@@ -5,8 +5,10 @@ from django.core.management.base import BaseCommand
 from openpyxl import Workbook
 from openpyxl.styles import Font
 
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import (
     ProtocoloPadraoDietaEspecial,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.terceirizada.models import Edital
