+++ sme_sigpae_api/dieta_especial/management/commands/fix_nomes_protocolos.py
@@ -1,7 +1,9 @@
 from django.core.management.base import BaseCommand
 
-from sme_sigpae_api.dieta_especial.models import (
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import (
     ProtocoloPadraoDietaEspecial,
+)
+from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     SolicitacaoDietaEspecial,
 )
