+++ sme_sigpae_api/processamento_arquivos/dieta_especial/__tests__/test_importa_alimentos.py
@@ -9,7 +9,7 @@ from sme_sigpae_api.dados_comuns.constants import StatusProcessamentoArquivo
 from sme_sigpae_api.dieta_especial.carga_dados.models import (
     ArquivoCargaAlimentosSubstitutos,
 )
-from sme_sigpae_api.dieta_especial.models import Alimento
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import Alimento
 from sme_sigpae_api.processamento_arquivos.dieta_especial import (
     importa_alimentos as imp_alimentos,
 )
