+++ sme_sigpae_api/processamento_arquivos/dieta_especial/__tests__/test_importa_dietas_especiais.py
@@ -4,12 +4,10 @@ from openpyxl.cell import Cell
 
 from sme_sigpae_api.dados_comuns.constants import StatusProcessamentoArquivo
 from sme_sigpae_api.dieta_especial.carga_dados.models import ArquivoCargaDietaEspecial
-from sme_sigpae_api.dieta_especial.models import (
-    SolicitacaoDietaEspecial,
-    SubstituicaoAlimento,
-)
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import SubstituicaoAlimento
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     ClassificacaoDieta,
+    SolicitacaoDietaEspecial,
 )
 from sme_sigpae_api.escola.models import Escola
 from sme_sigpae_api.processamento_arquivos.dieta_especial import (
