+++ sme_sigpae_api/produto/api/viewsets.py
@@ -68,7 +68,7 @@ from ...dados_comuns.permissions import (
     UsuarioTerceirizadaProduto,
 )
 from ...dados_comuns.utils import url_configs
-from ...dieta_especial.models import Alimento
+from ...dieta_especial.protocolo_padrao.models import Alimento
 from ...escola.models import DiretoriaRegional, Escola, Lote
 from ...relatorios.relatorios import (
     relatorio_produto_analise_sensorial,
