+++ utility/carga_dados/cardapio/_3_solicitacoes_dieta_especial.py
@@ -8,7 +8,10 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.db import transaction
 from faker import Faker
 
-from sme_sigpae_api.dieta_especial.models import Alimento, SubstituicaoAlimento
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import (
+    Alimento,
+    SubstituicaoAlimento,
+)
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     AlergiaIntolerancia,
     Anexo,
