+++ utility/carga_dados/dieta_especial/_13_dieta_especial.py
@@ -1,7 +1,7 @@
 import environ
 import pandas as pd
 
-from sme_sigpae_api.dieta_especial.models import Alimento
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import Alimento
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     AlergiaIntolerancia,
     ClassificacaoDieta,
