+++ utility/carga_dados/dieta_especial/importa_dados.py
@@ -9,7 +9,7 @@ from sme_sigpae_api.dieta_especial.data.motivo_alteracao_ue import (
     data_motivo_alteracao_ue,
 )
 from sme_sigpae_api.dieta_especial.data.motivo_negacao import data_motivo_negacoes
-from sme_sigpae_api.dieta_especial.models import Alimento
+from sme_sigpae_api.dieta_especial.protocolo_padrao.models import Alimento
 from sme_sigpae_api.dieta_especial.solicitacao_dieta_especial.models import (
     AlergiaIntolerancia,
     ClassificacaoDieta,
